requirements.txt
apturl==0.5.2
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==37.0.4
cs-bot
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
Flask==2.2.2
Flask-Cors==3.0.10
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
matchers==0.22
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
peewee==3.15.3
pexpect==4.8.0
Pillow==9.0.1
prometheus-client==0.14.1
protobuf==3.12.4
psutil==5.9.3
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
PyGObject==3.42.1
PyHamcrest==2.0.4
PyJWT==2.6.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.3.4
reportlab==3.6.8
requests==2.25.1
requests-unixsocket==0.2.0
schedule==1.1.0
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
systemd-python==234
typing_extensions==4.4.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==2.2.2
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
